# BackEnd COMSEG/Comseg.API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Comseg.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Pages/{id}": {
      "get": {
        "tags": [
          "Pages"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/PagesResponseBaseResponsePages"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/PagesResponseBaseResponsePages"
            }
          }
        }
      }
    },
    "/api/Roles": {
      "get": {
        "tags": [
          "Roles"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Role"
              }
            }
          }
        }
      }
    },
    "/api/Units": {
      "get": {
        "tags": [
          "Units"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Unit"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Units"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Unit"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Int32BaseResponseUnit"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/Int32BaseResponseUnit"
            }
          }
        }
      }
    },
    "/api/Units/{id}": {
      "put": {
        "tags": [
          "Units"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DtoUnit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BaseResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/BaseResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Units"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BaseResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/BaseResponse"
            }
          }
        }
      }
    },
    "/api/Users": {
      "post": {
        "tags": [
          "Users"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DtoLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UserLoginResponseBaseResponseLogin"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/UserLoginResponseBaseResponseLogin"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "BaseResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "listErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "DtoLogin": {
      "type": "object",
      "properties": {
        "userLogin": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "DtoUnit": {
      "required": [
        "abbreviation",
        "factor",
        "unitCode",
        "unitName",
        "unitStatus"
      ],
      "type": "object",
      "properties": {
        "modificationUser": {
          "type": "string"
        },
        "modificationDate": {
          "format": "date-time",
          "type": "string"
        },
        "unitCode": {
          "type": "string"
        },
        "unitName": {
          "type": "string"
        },
        "abbreviation": {
          "type": "string"
        },
        "factor": {
          "format": "int32",
          "type": "integer"
        },
        "unitStatus": {
          "type": "boolean"
        }
      }
    },
    "Int32BaseResponseUnit": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "listErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "responseResult": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "MenuResponse": {
      "type": "object",
      "properties": {
        "menuId": {
          "format": "int32",
          "type": "integer"
        },
        "menuName": {
          "type": "string"
        }
      }
    },
    "PagesResponse": {
      "type": "object",
      "properties": {
        "menuResponse": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/MenuResponse"
          }
        },
        "subMenuResponse": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SubMenuResponse"
          }
        }
      }
    },
    "PagesResponseBaseResponsePages": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "listErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "pageResponse": {
          "$ref": "#/definitions/PagesResponse"
        }
      }
    },
    "Role": {
      "type": "object",
      "properties": {
        "creationUser": {
          "type": "string"
        },
        "creationDate": {
          "format": "date-time",
          "type": "string"
        },
        "modificationUser": {
          "type": "string"
        },
        "modificationDate": {
          "format": "date-time",
          "type": "string"
        },
        "roleId": {
          "format": "int32",
          "type": "integer"
        },
        "roleName": {
          "type": "string"
        },
        "roleStatus": {
          "type": "boolean"
        },
        "rolePermissions": {
          "type": "string"
        }
      }
    },
    "SubMenuResponse": {
      "type": "object",
      "properties": {
        "menuId": {
          "format": "int32",
          "type": "integer"
        },
        "subMenuName": {
          "type": "string"
        },
        "subMenuPage": {
          "type": "string"
        },
        "subMenuStatus": {
          "type": "boolean"
        }
      }
    },
    "Unit": {
      "required": [
        "abbreviation",
        "factor",
        "unitCode",
        "unitName",
        "unitStatus"
      ],
      "type": "object",
      "properties": {
        "creationUser": {
          "type": "string"
        },
        "creationDate": {
          "format": "date-time",
          "type": "string"
        },
        "modificationUser": {
          "type": "string"
        },
        "modificationDate": {
          "format": "date-time",
          "type": "string"
        },
        "unitId": {
          "format": "int32",
          "type": "integer"
        },
        "unitCode": {
          "type": "string"
        },
        "unitName": {
          "type": "string"
        },
        "abbreviation": {
          "type": "string"
        },
        "factor": {
          "format": "int32",
          "type": "integer"
        },
        "unitStatus": {
          "type": "boolean"
        }
      }
    },
    "UserLoginResponse": {
      "type": "object",
      "properties": {
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "userName": {
          "type": "string"
        },
        "roleId": {
          "format": "int32",
          "type": "integer"
        },
        "token": {
          "type": "string"
        }
      }
    },
    "UserLoginResponseBaseResponseLogin": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "listErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "userResponse": {
          "$ref": "#/definitions/UserLoginResponse"
        }
      }
    }
  }
}